+++ build.py
@@ -10,11 +10,12 @@ import os
 from setuptools import Distribution
 from setuptools.command.build_ext import build_ext
 
+
 def build() -> None:
     ext_modules = get_extensions()
 
     distribution = Distribution({
-        "name": "package",
+        "name": "lco-banzai",
         "ext_modules": ext_modules
     })
     cmd = build_ext(distribution)